+++ app-watch/build.gradle
@@ -5,12 +5,12 @@ plugins {
 
 android {
     namespace 'io.foundy.hanstargramwatch'
-    compileSdk 32
+    compileSdk 33
 
     defaultConfig {
         applicationId "io.foundy.hanstargramwatch"
         minSdk 26
-        targetSdk 32
+        targetSdk 33
         versionCode 1
         versionName "1.0"
 
@@ -28,6 +28,8 @@ android {
 }
 
 dependencies {
+    implementation project(":data")
+    implementation project(":domain")
 
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'com.google.android.gms:play-services-wearable:18.0.0'
