+++ app/build.gradle
@@ -38,6 +38,9 @@ android {
 }
 
 dependencies {
+    implementation project(":data")
+    implementation project(":domain")
+    implementation project(':common')
 
     // Glide
     implementation "com.github.bumptech.glide:glide:$glide"
