--- app/src/main/java/io/foundy/hanstargram/repository/model/FollowDto.kt
@@ -1,7 +1,0 @@
-package io.foundy.hanstargram.repository.model
-
-data class FollowDto(
-    val uuid: String = "",
-    val followerUuid: String = "",
-    val followeeUuid: String = ""
-)
