--- app/src/main/java/io/foundy/hanstargram/repository/model/LikeDto.kt
@@ -1,7 +1,0 @@
-package io.foundy.hanstargram.repository.model
-
-data class LikeDto(
-    val uuid: String = "",
-    val userUuid: String = "",
-    val postUuid: String = ""
-)
