--- app/src/main/java/io/foundy/hanstargram/repository/model/PostDto.kt
@@ -1,11 +1,0 @@
-package io.foundy.hanstargram.repository.model
-
-import java.util.*
-
-data class PostDto(
-    val uuid: String = "",
-    val writerUuid: String = "",
-    val content: String = "",
-    val dateTime: Date = Date(),
-    val imageUrl: String = ""
-)
