+++ common/src/main/java/io/foundy/common/base/ViewBindingFragment.kt
@@ -1,0 +1,33 @@
+package io.foundy.common.base
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.annotation.CallSuper
+import androidx.fragment.app.Fragment
+import androidx.viewbinding.ViewBinding
+
+abstract class ViewBindingFragment<VB : ViewBinding> : Fragment() {
+
+    protected abstract val bindingInflater: (LayoutInflater, ViewGroup?, Boolean) -> VB
+
+    private var _binding: VB? = null
+    protected val binding: VB get() = requireNotNull(_binding)
+
+    @CallSuper
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = bindingInflater.invoke(inflater, container, false)
+        return binding.root
+    }
+
+    @CallSuper
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
