+++ data/src/main/java/io/foundy/data/model/FollowDto.kt
@@ -1,0 +1,7 @@
+package io.foundy.data.model
+
+data class FollowDto(
+    val uuid: String = "",
+    val followerUuid: String = "",
+    val followeeUuid: String = ""
+)
