+++ data/src/main/java/io/foundy/data/model/LikeDto.kt
@@ -1,0 +1,7 @@
+package io.foundy.data.model
+
+data class LikeDto(
+    val uuid: String = "",
+    val userUuid: String = "",
+    val postUuid: String = ""
+)
