+++ data/src/main/java/io/foundy/data/model/UserDto.kt
@@ -1,0 +1,9 @@
+package io.foundy.data.model
+
+data class UserDto(
+    val uuid: String = "",
+    val name: String = "",
+    val email: String? = null,
+    val password: String? = null,
+    val profileImageUrl: String? = null
+)
