+++ data/src/main/java/io/foundy/data/source/UserPagingSource.kt
@@ -1,0 +1,44 @@
+package io.foundy.data.source
+
+import androidx.paging.PagingSource
+import androidx.paging.PagingState
+import com.google.firebase.firestore.Query
+import com.google.firebase.firestore.QuerySnapshot
+import com.google.firebase.firestore.ktx.toObjects
+import io.foundy.data.model.UserDto
+import kotlinx.coroutines.tasks.await
+
+class UserPagingSource(
+    private val queryUsersByName: Query
+) : PagingSource<QuerySnapshot, UserDto>() {
+
+    override fun getRefreshKey(state: PagingState<QuerySnapshot, UserDto>): QuerySnapshot? {
+        return null
+    }
+
+    override suspend fun load(
+        params: LoadParams<QuerySnapshot>
+    ): LoadResult<QuerySnapshot, UserDto> {
+        return try {
+            val currentPage = params.key ?: queryUsersByName.get().await()
+            if (currentPage.isEmpty) {
+                return LoadResult.Page(
+                    data = emptyList(),
+                    prevKey = null,
+                    nextKey = null
+                )
+            }
+            val lastVisiblePost = currentPage.documents[currentPage.size() - 1]
+            val nextPage = queryUsersByName.startAfter(lastVisiblePost).get().await()
+            val userDto = currentPage.toObjects<UserDto>()
+            println(userDto)
+            LoadResult.Page(
+                data = userDto,
+                prevKey = null,
+                nextKey = nextPage
+            )
+        } catch (e: Exception) {
+            LoadResult.Error(e)
+        }
+    }
+}
