+++ domain/src/main/java/io/foundy/domain/model/Post.kt
@@ -1,0 +1,13 @@
+package io.foundy.domain.model
+
+data class Post(
+    val uuid: String,
+    val writerName: String,
+    val writerProfileImageUrl: String?,
+    val content: String,
+    val imageUrl: String,
+    val likeCount: Int,
+    val meLiked: Boolean,
+    val isMine: Boolean,
+    val timeAgo: String
+)
