+++ settings.gradle
@@ -15,3 +15,6 @@ dependencyResolutionManagement {
 rootProject.name = "Hanstargram"
 include ':app'
 include ':app-watch'
+include ':data'
+include ':domain'
+include ':common'
